feat: Separate logics in files.
