+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@adempiere/grpc-core-client",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "ADempiere Core Client with base functionality for ADempiere client write in Javascript for gRPC service",
   "main": "index.js",
   "scripts": {
@@ -12,6 +12,12 @@
   },
   "author": "Yamel Senih",
   "license": "GPL-3.0-or-later",
+  "contributors": [
+    {
+      "name": "Edwin Betancourt",
+      "url": "https://github.com/EdwinBetanc0urt/"
+    }
+  ],
   "bugs": {
     "url": "https://github.com/erpcya/gRPC-Core-Client/issues"
   },
